+++ src/js/renderers/dailymotion.js
@@ -432,6 +432,9 @@ const DailyMotionIframeRenderer = {
 		dmSettings.params = Object.assign({}, dm.options.dailymotion);
 
 		// Check for `autoplay` and `muted` attributes to override settings
+		if (mediaElement.originalNode.controls) {
+			dmSettings.params.controls = true;
+		}
 		if (mediaElement.originalNode.autoplay) {
 			dmSettings.params.autoplay = true;
 		}